Split out BasicIT.java
